fix(ProgressBar): use int in place of std::size_t

+++ include/ProgressBar.hpp
@@ -61,7 +61,7 @@ public:
 
 private:
     std::string writeTime() const;
-    std::string drawBar(const size_t uWidth) const;
+    std::string drawBar(const int uWidth) const;
     std::string drawProgress() const;
     std::string drawPrefix() const;
